refactor: move js-files to src directory
