+++ src/color-format.js
